Modify ProtobufMessageConverter to use DynamicMessage and Descriptors
Now, the ProtobufMessageConverter uses the proto file descriptor to
encode what type of message it's converting and uses that to decode any
message it receives.  This should make it more platform independent and
we don't need to create a new message converter for each type of message
that we are sending.